mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerExecution.java
package com.interface21.webmvc.servlet.mvc.tobe;

import java.lang.reflect.Method;

import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import com.interface21.webmvc.servlet.ModelAndView;

public class HandlerExecution {

    private final Class<?> controllerClass;
    private final Method method;

    public HandlerExecution(final Class<?> controllerClass, final Method method) {
        this.controllerClass = controllerClass;
        this.method = method;
    }

    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
        Object controller = controllerClass.getDeclaredConstructor().newInstance();
        return (ModelAndView) method.invoke(controller, request, response);
    }
}
